Refactor ShardSelector (#118)
* Renamed __cortex_shard__ label to __query_shard__


* Renamed ShardSelector.String() to ShardSelector.LabelValue()


* Change ShardIndex and ShardCount to uint64 to play it nicer with blocks storage


* Fixed TestNewMockShardedqueryable


* Fixed parseShard() and moved tests to the right package


* Fixed unit tests after rebase

+++ pkg/storegateway/bucket.go
@@ -674,7 +674,7 @@ func blockSeries(
 				seriesCacheStats.seriesHashCacheHits++
 			}
 
-			if hash%uint64(shard.ShardCount) != uint64(shard.ShardIndex) {
+			if hash%shard.ShardCount != shard.ShardIndex {
 				continue
 			}
 		}
